requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
python-decouple==3.8
psycopg2==2.9.5